Fixing the dependency issue
Problem was that silver used a different scalatest version. Setting the
dependency version in the silver build file to provided did the trick.
The tests now work both in intellij and with sbt out of the box. The
assembly also works.
Additionally the tests are now skipped during the assembly phase.